fix first lab in accordinates with contract and remove depricated struct with functions (used for add init extensions)
