+++ src/backend/boss_operations/boss_operations.c
@@ -1,8 +1,9 @@
-#include "../../include/memory/memcache_map.h"
-#include "../../include/boss_operations/boss_operations.h"
-#include "../../include/memory/cache.h"
-#include "../../include/logger/logger.h"
-#include "../../include/guc/guc.h"
+#include "memory/memcache_map.h"
+#include "boss_operations.h"
+#include "master.h"
+#include "memory/cache.h"
+#include "logger/logger.h"
+#include "guc/guc.h"
 #include <stdlib.h>
 #include <assert.h>
 #include <string.h>
@@ -18,18 +19,18 @@ static inline Op_stack_ptr create_stack()
 }
 
 static void push_to_stack(Op_stack_ptr *stack, Boss_op_name op_name, void *operation_data)
-{   
-    Op_stack_elem *new_element = (Op_stack_elem *) malloc(sizeof(Op_stack_elem));
+{
+    Op_stack_elem *new_element = (Op_stack_elem *)malloc(sizeof(Op_stack_elem));
     assert(new_element != NULL);
-    
+
     new_element->op_name = op_name;
     new_element->operation_data = operation_data;
     new_element->next_elem = *stack;
     *stack = new_element;
 }
 
-static void* pop_from_stack(Op_stack_ptr *stack, Boss_op_name *op_name)
-{   
+static void *pop_from_stack(Op_stack_ptr *stack, Boss_op_name *op_name)
+{
     *op_name = (*stack)->op_name;
     void *operation_data = (*stack)->operation_data;
 
@@ -39,24 +40,16 @@ static void* pop_from_stack(Op_stack_ptr *stack, Boss_op_name *op_name)
     return operation_data;
 }
 
-//============operations for shared libraries=============
+//============operations for shared libraries (hooks)=============
 
-void init_boss_op()
-{
-    boss_op = create_stack();
-}
-
-int cache_write_op
-(
-    const char *key, 
+int cache_write_op(
+    const char *key,
     const char *message,
     const size_t message_length,
-    const unsigned TTL
-)
+    const unsigned TTL)
 {
     if (strlen(key) + 1 > MAX_KEY_SIZE)
     {
-        printf("Here\n");
         elog(ERROR, "Size of arguments too large");
         return -1;
     }
@@ -74,8 +67,8 @@ void print_cache_op(const char *key)
         elog(ERROR, "Size of arguments too large");
         return;
     }
-    
-    char *message = (char *) cache_read(key, NULL, 0);
+
+    char *message = (char *)cache_read(key, NULL, 0);
     printf("%s\n", message);
 }
 
@@ -87,13 +80,13 @@ void register_background_worker(char *callback_name, char *bg_worker_name, bool 
         return;
     }
 
-    BGWorker_data *bg_worker_data = (BGWorker_data *) malloc(sizeof(BGWorker_data));
+    BGWorker_data *bg_worker_data = (BGWorker_data *)malloc(sizeof(BGWorker_data));
     assert(bg_worker_data);
 
-    bg_worker_data->callback_name = (char *) malloc(strlen(callback_name) + 1);
+    bg_worker_data->callback_name = (char *)malloc(strlen(callback_name) + 1);
     strcpy(bg_worker_data->callback_name, callback_name);
 
-    bg_worker_data->bg_worker_name = (char *) malloc(strlen(bg_worker_name) + 1);
+    bg_worker_data->bg_worker_name = (char *)malloc(strlen(bg_worker_name) + 1);
     strcpy(bg_worker_data->bg_worker_name, bg_worker_name);
 
     bg_worker_data->need_observer = need_observer;
@@ -120,11 +113,9 @@ void define_custom_long_variable_op(char *name, const char *descr, const long bo
 
 void define_custom_string_variable_op(char *name, const char *descr, const char *boot_value, const int8_t context)
 {
-    if 
-    (
+    if (
         strlen(name) + 1 > MAX_NAME_LENGTH ||
-        strlen(descr) + 1 > MAX_DESCRIPTION_LENGTH
-    )
+        strlen(descr) + 1 > MAX_DESCRIPTION_LENGTH)
     {
         elog(ERROR, "Size of arguments too large");
         return;
@@ -152,7 +143,8 @@ char *get_config_string_parameter_op(const char *name, const int8_t context)
     {
         elog(ERROR, "Varibale with name %s not exists in this context\n", name);
         return NULL;
-    } else if (vartype != STRING)
+    }
+    else if (vartype != STRING)
     {
         elog(ERROR, "Varibale with name %s isn`t string\n", name);
         return NULL;
@@ -163,7 +155,12 @@ char *get_config_string_parameter_op(const char *name, const int8_t context)
     return var.str;
 }
 
-//============operations for boss=============
+//============operations for boss (boss_operations)=============
+
+inline void init_boss_op()
+{
+    boss_op = create_stack();
+}
 
 inline void clear_bg_worker_info(BGWorker_data *bg_worker_data)
 {
