+++ src/backend/shlib_operations/README.md
@@ -12,6 +12,6 @@
     extern void init(void *arg, ...).
 4. Если расширению требуется выполнение некоторый операций от основного процесса (например создание отдельного процесса
 для себя), то оно может воспользоваться одим из указателей на функцию из структуры Boss_op_func
-(backend/include/boss_operations/boss_operations.h), указатель на которую передаётся в расширение первым аргументом.
+(backend/include/boss_operations.h), указатель на которую передаётся в расширение первым аргументом.
 Указатели на функции валидны только в рамках выполнения функции init.
 P.S. В дальнейшем протокол взаимодействия основоной программы с раширением возможно будет расширен.
