+++ src/backend/shlib_operations/initializer.c
@@ -7,8 +7,8 @@
 #include "utils/stack.h"
 #include "shlib_operations/operations.h"
 #include "logger/logger.h"
-#include "boss_operations/boss_operations.h"
-#include "boss_operations/hook.h"
+#include "boss_operations.h"
+#include "master.h"
 
 /**
     \brief Initialize all extensions (call init function with some args)
@@ -21,14 +21,6 @@ void init_all_exetensions()
 
     Stack_ptr curr_stack = lib_stack;
 
-    Boss_op_func *op_func = malloc(sizeof(Boss_op_func));
-    op_func->cache_write_op = cache_write_op;
-    op_func->print_cache_op = print_cache_op;
-    op_func->register_background_worker = register_background_worker;
-    op_func->define_custom_long_variable_op = define_custom_long_variable_op;
-    op_func->define_custom_string_variable_op = define_custom_string_variable_op;
-    op_func->get_config_string_parameter_op = get_config_string_parameter_op;
-
     for (size_t i = 0; i < get_stack_size(lib_stack); i++)
     {
         void (*function)() = dlsym(curr_stack->data, "init");
@@ -38,12 +30,11 @@ void init_all_exetensions()
             curr_stack = curr_stack->next_elem;
             continue;
         }
-        function(op_func);
+        function();
         exec_requests(curr_stack->data);
 
         curr_stack = curr_stack->next_elem;
     }
-    free(op_func);
 
     if (executor_start_hook) {
         executor_start_hook();
